# nomlog-api/swagger.yaml
openapi: 3.0.3
info:
  title: Nomlog API
  description: API for meal logging and nutrition tracking
  version: 1.0.0
  contact:
    name: Nomlog API Support
servers:
  - url: http://localhost:3001/api/v1
    description: Local development server
  - url: https://nomlog-api.onrender.com/api/v1
    description: Production server
...
paths:
...
  /water:
    get:
      tags:
        - Water
      summary: Get water log for a date
      description: Retrieve water intake log for a specific date
      operationId: getWaterLog
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Date in YYYY-MM-DD format
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}-\d{2}$'
            example: 2024-03-15
      responses:
        '200':
          description: Water log retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  waterLog:
                    oneOf:
                      - $ref: '#/components/schemas/WaterLog'
                      - type: 'null'
                    description: Water log for the date, or null if none exists
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'

    put:
      tags:
        - Water
      summary: Update water log for a date
      description: Create or update water intake log for a specific date (upsert)
      operationId: updateWaterLog
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - date
                - glasses
              properties:
                date:
                  type: string
                  pattern: '^\d{4}-\d{2}-\d{2}$'
                  description: Date in YYYY-MM-DD format
                  example: 2024-03-15
                glasses:
                  type: integer
                  minimum: 0
                  description: Total number of glasses of water
                  example: 4
            examples:
              updateWater:
                summary: Update water intake
                value:
                  date: 2024-03-15
                  glasses: 4
      responses:
        '200':
          description: Water log updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Water log updated successfully
                  waterLog:
                    $ref: '#/components/schemas/WaterLog'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    WaterLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        date:
          type: string
          pattern: '^\d{4}-\d{2}-\d{2}$'
          example: 2024-03-15
          description: Date in YYYY-MM-DD format
        glasses:
          type: integer
          minimum: 0
          example: 4
          description: Total number of glasses of water
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          example: Error message
        message:
          type: string
          example: Detailed error message
        details:
          type: array
          items:
            type: object
          description: Validation error details (for 400 responses)
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'